restructuring and reformatting of source code
