+++ examples/misc/simple.lp
@@ -1,0 +1,4 @@
+a(1..5).
+b(5..10).
+
+:- a(X), b(X).
