+++ examples/misc/soup.lp
@@ -1,0 +1,3 @@
+#include "test/blob.lp".
+
+soup(10).
