+++ examples/misc/test/blob.lp
@@ -1,0 +1,3 @@
+#include "../test_inert.lp".
+
+blob(123).
