+++ examples/misc/test_inert.lp
@@ -1,0 +1,1 @@
+blub(42).
