+++ src/clingexplaid/muc/__init__.py
@@ -1,0 +1,5 @@
+"""
+Minimal Unsatisfiable Core Utilities
+"""
+
+from .core_computer import CoreComputer
