+++ src/clingexplaid/muc/core_computer.py
@@ -1,12 +1,9 @@
-"""
-Unsatisfiable Core Utilities
-"""
 from typing import Optional, Set, Tuple
 from itertools import chain, combinations
 
 import clingo
 
-from . import Assumption, AssumptionSet, SymbolSet, get_solver_literal_lookup
+from ..utils import Assumption, AssumptionSet, SymbolSet, get_solver_literal_lookup
 
 
 class CoreComputer:
@@ -21,9 +18,7 @@ class CoreComputer:
         self.literal_lookup = get_solver_literal_lookup(control=self.control)
         self.minimal: Optional[AssumptionSet] = None
 
-    def _solve(
-        self, assumptions: Optional[AssumptionSet] = None
-    ) -> Tuple[bool, SymbolSet, SymbolSet]:
+    def _solve(self, assumptions: Optional[AssumptionSet] = None) -> Tuple[bool, SymbolSet, SymbolSet]:
         """
         Internal function that is used to make the single solver calls for finding the minimal unsatisfiable core.
         """
@@ -32,23 +27,15 @@ class CoreComputer:
 
         with self.control.solve(assumptions=list(assumptions), yield_=True) as solve_handle:  # type: ignore[union-attr]
             satisfiable = bool(solve_handle.get().satisfiable)
-            model = (
-                solve_handle.model().symbols(atoms=True)
-                if solve_handle.model() is not None
-                else []
-            )
-            core = {
-                self.literal_lookup[literal_id] for literal_id in solve_handle.core()
-            }
+            model = solve_handle.model().symbols(atoms=True) if solve_handle.model() is not None else []
+            core = {self.literal_lookup[literal_id] for literal_id in solve_handle.core()}
 
         return satisfiable, set(model), core
 
-    def _compute_single_minimal(
-        self, assumptions: Optional[AssumptionSet] = None
-    ) -> AssumptionSet:
+    def _compute_single_minimal(self, assumptions: Optional[AssumptionSet] = None) -> AssumptionSet:
         """
         Function to compute a single minimal unsatisfiable core from the passed set of assumptions and the program of
-        the CoreComputer. If there is not minimal unsatisfiable core, since for example the program with assumptions
+        the CoreComputer. If there is no minimal unsatisfiable core, since for example the program with assumptions
         assumed is satisfiable, an empty set is returned. The algorithm that is used to compute this minimal
         unsatisfiable core is the iterative deletion algorithm.
         """
@@ -57,9 +44,7 @@ class CoreComputer:
 
         # check that the assumption set isn't empty
         if not assumptions:
-            raise ValueError(
-                "A minimal unsatisfiable core cannot be computed on an empty assumption set"
-            )
+            raise ValueError("A minimal unsatisfiable core cannot be computed on an empty assumption set")
 
         # check if the problem with the full assumption set is unsatisfiable in the first place, and if not skip the
         # rest of the algorithm and return an empty set.
@@ -101,8 +86,7 @@ class CoreComputer:
         """
         assumptions = self.assumption_set
         assumption_powerset = chain.from_iterable(
-            combinations(assumptions, r)
-            for r in reversed(range(len(list(assumptions)) + 1))
+            combinations(assumptions, r) for r in reversed(range(len(list(assumptions)) + 1))
         )
 
         found_sat = []
@@ -133,8 +117,3 @@ class CoreComputer:
                 # if the maximum muc amount is found stop search
                 if max_mucs is not None and len(found_mucs) == max_mucs:
                     break
-
-
-__all__ = [
-    CoreComputer.__name__,
-]
