+++ src/clingexplaid/propagators/__init__.py
@@ -1,0 +1,10 @@
+"""
+Propagators for Explanation
+"""
+
+from typing import List
+
+from .propagator_decision_order import DecisionOrderPropagator
+
+DecisionLevel = List[int]
+DecisionLevelList = List[DecisionLevel]
