+++ src/clingexplaid/transformers/constants.py
@@ -1,0 +1,2 @@
+REMOVED_TOKEN = "__REMOVED__"
+RULE_ID_SIGNATURE = "_rule"
