+++ src/clingexplaid/unsat_constraints/__init__.py
@@ -1,0 +1,5 @@
+"""
+Functionality for Unsat Constraints
+"""
+
+from .unsat_constraint_computer import UnsatConstraintComputer
