+++ src/clingexplaid/unsat_constraints/constants.py
@@ -1,0 +1,1 @@
+UNSAT_CONSTRAINT_SIGNATURE = "__unsat__"
